sdelal_Thread_priority

+++ lab-client/src/main/java/ProcessEngine/GraphicCore/MainWindow/DataSheet/SorteByColumn/SortByKey.java
@@ -8,7 +8,7 @@ public class SortByKey {
 
     public static Vector<String[]> sortByKeyAscendingOrder(Vector<String[]> arr) {
         arr = arr.stream()
-            .sorted((a, b) -> Integer.compare(Integer.parseInt(a[0]), Integer.parseInt(b[0])))
+            .sorted((a, b) -> Long.compare(Long.parseLong(a[0]), Long.parseLong(b[0])))
             .collect(Collectors.toCollection(Vector::new));
         return arr;
     }
